refactor: Rename controllers package to api/user
