+++ api/user/user.go
@@ -1,4 +1,4 @@
-package controllers
+package user
 
 import (
 	"errors"
